+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * **Temporarily** use a simple textbox instead of `<AutoSuggest>` in generated query forms. Addresses UILDP-147.
 * Closing the results list from a "Run report" query no longer clears entered parameters. Fixes UILDP-138.
 * Support GitLab as well as GitHub as a source for templated reports. Fixes UILDP-132.
+* When there are no tables in the selected schema (e.g. because a back-end problems reports no tables at all), the UI is now robust against this condition. Fixes UILDP-151.
 
 ## [2.1.0](https://github.com/folio-org/ui-ldp/tree/v2.1.0) (2024-03-19)